modification affichage : - la moyenne  n'affiche plus les decimales. - changement de la police de la partie actualiter de la partie pour etre compatible avec la raspberry - ajout de fleche pour mettre en avant la possibiliter de finir la partie - modification de son de l'ihm

+++ trunk/ecran-DARTS/joueur.cpp
@@ -76,7 +76,7 @@ int Joueur::getFlechette() const
  * @fn Joueur::getFlechette
  * @return float
  */
-float Joueur::getMoyenneVolee() const
+int Joueur::getMoyenneVolee() const
 {
     return this->moyenneVolee;
 }
@@ -120,7 +120,7 @@ void Joueur::setEliminer(bool elimine)
  * @fn Joueur::setMoyenneVolee
  * @param moyenneVolee
  */
-void Joueur::setMoyenneVolee(float moyenneVolee)
+void Joueur::setMoyenneVolee(int moyenneVolee)
 {
     this->moyenneVolee = moyenneVolee;
 }
